dev-requirements.txt
pip
pylint
flake8
bandit
mypy
wheel
tox
coverage
pytest-cov
